@@ -1,25 +1,19 @@
 import json
+from pathlib import Path
 
-from voicevox_engine.dev.core import mock as core
-from voicevox_engine.dev.synthesis_engine.mock import MockSynthesisEngine
+from voicevox_engine.dev.core import MockCoreWrapper
+from voicevox_engine.dev.tts_engine.mock import MockTTSEngine
 from voicevox_engine.preset import PresetManager
 from voicevox_engine.setting import USER_SETTING_PATH, SettingLoader
+from voicevox_engine.tts_pipeline.tts_engine import CoreAdapter
 from voicevox_engine.utility import engine_root
 
-if __name__ == "__main__":
-    import run
 
-    app = run.generate_app(
-        synthesis_engines={"mock": MockSynthesisEngine(speakers=core.metas())},
-        latest_core_version="mock",
-        setting_loader=SettingLoader(USER_SETTING_PATH),
-        preset_manager=PresetManager(  # FIXME: impl MockPresetManager
-            preset_path=engine_root() / "presets.yaml",
-        ),
-    )
-    with open("docs/api/index.html", "w") as f:
-        f.write(
-            """<!DOCTYPE html>
+def generate_api_docs_html(schema: str) -> str:
+    """OpenAPI schema から API ドキュメント HTML を生成する"""
+
+    return (
+        """<!DOCTYPE html>
 <html lang="ja">
 <head>
     <title>voicevox_engine API Document</title>
@@ -34,5 +28,31 @@ if __name__ == "__main__":
     </script>
 </body>
 </html>"""
-            % json.dumps(app.openapi())
-        )
+        % schema
+    )
+
+
+if __name__ == "__main__":
+    import run
+
+    mock_core = MockCoreWrapper()
+    # FastAPI の機能を用いて OpenAPI schema を生成する
+    app = run.generate_app(
+        tts_engines={"mock": MockTTSEngine()},
+        cores={"mock": CoreAdapter(mock_core)},
+        latest_core_version="mock",
+        setting_loader=SettingLoader(USER_SETTING_PATH),
+        preset_manager=PresetManager(  # FIXME: impl MockPresetManager
+            preset_path=engine_root() / "presets.yaml",
+        ),
+    )
+    api_schema = json.dumps(app.openapi())
+
+    # API ドキュメント HTML を生成する
+    api_docs_html = generate_api_docs_html(api_schema)
+
+    # HTML ファイルとして保存する
+    api_docs_root = Path("docs/api")  # 'upload-docs' workflow の対象
+    output_path = api_docs_root / "index.html"
+    output_path.parent.mkdir(parents=True, exist_ok=True)
+    output_path.write_text(api_docs_html)
