@@ -1,20 +1,20 @@
 from unittest import TestCase
 
-from voicevox_engine.mora_list import openjtalk_mora2text
+from voicevox_engine.tts_pipeline.mora_list import mora_phonemes_to_mora_kana
 
 
 class TestOpenJTalkMoraList(TestCase):
     def test_mora2text(self):
-        self.assertEqual("ッ", openjtalk_mora2text["cl"])
-        self.assertEqual("ティ", openjtalk_mora2text["ti"])
-        self.assertEqual("トゥ", openjtalk_mora2text["tu"])
-        self.assertEqual("ディ", openjtalk_mora2text["di"])
+        self.assertEqual("ッ", mora_phonemes_to_mora_kana["cl"])
+        self.assertEqual("ティ", mora_phonemes_to_mora_kana["ti"])
+        self.assertEqual("トゥ", mora_phonemes_to_mora_kana["tu"])
+        self.assertEqual("ディ", mora_phonemes_to_mora_kana["di"])
         # GitHub issue #60
-        self.assertEqual("ギェ", openjtalk_mora2text["gye"])
-        self.assertEqual("イェ", openjtalk_mora2text["ye"])
+        self.assertEqual("ギェ", mora_phonemes_to_mora_kana["gye"])
+        self.assertEqual("イェ", mora_phonemes_to_mora_kana["ye"])
 
     def test_mora2text_injective(self):
         """異なるモーラが同じ読みがなに対応しないか確認する"""
-        values = list(openjtalk_mora2text.values())
+        values = list(mora_phonemes_to_mora_kana.values())
         uniq_values = list(set(values))
         self.assertCountEqual(values, uniq_values)
