@@ -1,118 +1,148 @@
 import json
-from logging import getLogger
-from typing import Any, Dict, List
+from pathlib import Path
 
 import numpy as np
-from pyopenjtalk import tts
-from soxr import resample
+from numpy import ndarray
+from numpy.typing import NDArray
 
-DUMMY_TEXT = "これはダミーのテキストです"
-
-
-def initialize(path: str, use_gpu: bool, *args: List[Any]) -> None:
-    pass
+from ...core_wrapper import CoreWrapper
 
 
-def yukarin_s_forward(length: int, **kwargs: Dict[str, Any]) -> np.ndarray:
-    logger = getLogger("uvicorn")  # FastAPI / Uvicorn 内からの利用のため
-    logger.info(
-        "Sorry, yukarin_s_forward() is a mock. Return values are incorrect.",
-    )
-    return np.ones(length) / 5
+class MockCoreWrapper(CoreWrapper):
+    """`CoreWrapper` Mock"""
 
+    def __init__(
+        self,
+        use_gpu: bool = False,
+        core_dir: Path | None = None,
+        cpu_num_threads: int = 0,
+        load_all_models: bool = False,
+    ) -> None:
+        self.default_sampling_rate = 24000
 
-def yukarin_sa_forward(length: int, **kwargs: Dict[str, Any]) -> np.ndarray:
-    logger = getLogger("uvicorn")  # FastAPI / Uvicorn 内からの利用のため
-    logger.info(
-        "Sorry, yukarin_sa_forward() is a mock. Return values are incorrect.",
-    )
-    return np.ones((1, length)) * 5
+    def metas(self) -> str:
+        return json.dumps(
+            [
+                {
+                    "name": "dummy1",
+                    "styles": [
+                        {"name": "style0", "id": 0},
+                        {"name": "style1", "id": 2},
+                        {"name": "style2", "id": 4},
+                        {"name": "style3", "id": 6},
+                    ],
+                    "speaker_uuid": "7ffcb7ce-00ec-4bdc-82cd-45a8889e43ff",
+                    "version": "mock",
+                },
+                {
+                    "name": "dummy2",
+                    "styles": [
+                        {"name": "style0", "id": 1},
+                        {"name": "style1", "id": 3},
+                        {"name": "style2", "id": 5},
+                        {"name": "style3", "id": 7},
+                    ],
+                    "speaker_uuid": "388f246b-8c41-4ac1-8e2d-5d79f3ff56d9",
+                    "version": "mock",
+                },
+                {
+                    "name": "dummy3",
+                    "styles": [
+                        {"name": "style0", "id": 8},
+                    ],
+                    "speaker_uuid": "35b2c544-660e-401e-b503-0e14c635303a",
+                    "version": "mock",
+                },
+                {
+                    "name": "dummy4",
+                    "styles": [
+                        {"name": "style0", "id": 9},
+                    ],
+                    "speaker_uuid": "b1a81618-b27b-40d2-b0ea-27a9ad408c4b",
+                    "version": "mock",
+                },
+            ]
+        )
 
+    def yukarin_s_forward(
+        self, length: int, phoneme_list: ndarray, style_id: ndarray
+    ) -> NDArray[np.floating]:
+        """音素系列サイズ・音素ID系列・スタイルIDから音素長系列を生成する"""
+        result = []
+        # mockとしての適当な処理、特に意味はない
+        for i in range(length):
+            result.append(round((phoneme_list[i] * 0.0625 + style_id).item(), 2))
+        return np.array(result)
 
-def decode_forward(length: int, **kwargs: Dict[str, Any]) -> np.ndarray:
-    """
-    合成音声の波形データをNumPy配列で返します。ただし、常に固定の文言を読み上げます（DUMMY_TEXT）
-    参照→SynthesisEngine のdocstring [Mock]
+    def yukarin_sa_forward(
+        self,
+        length: int,
+        vowel_phoneme_list: ndarray,
+        consonant_phoneme_list: ndarray,
+        start_accent_list: ndarray,
+        end_accent_list: ndarray,
+        start_accent_phrase_list: ndarray,
+        end_accent_phrase_list: ndarray,
+        style_id: ndarray,
+    ) -> NDArray[np.floating]:
+        """モーラ系列サイズ・母音系列・子音系列・アクセント位置・アクセント句区切り・スタイルIDからモーラ音高系列を生成する"""
+        assert length > 1, "前後無音を必ず付与しなければならない"
 
-    Parameters
-    ----------
-    length : int
-        フレームの長さ
+        result = []
+        # mockとしての適当な処理、特に意味はない
+        for i in range(length):
+            result.append(
+                round(
+                    (
+                        (
+                            vowel_phoneme_list[0][i]
+                            + consonant_phoneme_list[0][i]
+                            + start_accent_list[0][i]
+                            + end_accent_list[0][i]
+                            + start_accent_phrase_list[0][i]
+                            + end_accent_phrase_list[0][i]
+                        )
+                        * 0.0625
+                        + style_id
+                    ).item(),
+                    2,
+                )
+            )
+        return np.array(result)[np.newaxis]
 
-    Returns
-    -------
-    wave : np.ndarray
-        音声合成した波形データ
+    def decode_forward(
+        self,
+        length: int,
+        phoneme_size: int,
+        f0: ndarray,
+        phoneme: ndarray,
+        style_id: ndarray,
+    ) -> NDArray[np.floating]:
+        """フレーム長・音素種類数・フレーム音高・フレーム音素onehot・スタイルIDからダミー音声波形を生成する"""
+        # 入力値を反映し、長さが 256 倍であるダミー配列を出力する
+        result: list[ndarray] = []
+        for i in range(length):
+            result += [
+                (f0[i, 0] * (np.where(phoneme[i] == 1)[0] / phoneme_size) + style_id)
+            ] * 256
+        return np.array(result)
 
-    Note
-    -------
-        ここで行う音声合成では、調声（ピッチ等）を反映しない
-        また、入力内容によらず常に固定の文言を読み上げる
-
-        # pyopenjtalk.tts()の出力仕様
-        dtype=np.float64, 16 bit, mono 48000 Hz
-
-        # resampleの説明
-        非モックdecode_forwardと合わせるために、出力を24kHzに変換した。
-    """
-    logger = getLogger("uvicorn")  # FastAPI / Uvicorn 内からの利用のため
-    logger.info(
-        "Sorry, decode_forward() is a mock. Return values are incorrect.",
-    )
-    wave, sr = tts(DUMMY_TEXT)
-    wave = resample(wave.astype("int16"), 48000, 24000)
-    return wave
-
-
-def metas() -> str:
-    return json.dumps(
-        [
+    def supported_devices(self):
+        return json.dumps(
             {
-                "name": "dummy1",
-                "styles": [
-                    {"name": "style0", "id": 0},
-                    {"name": "style1", "id": 2},
-                    {"name": "style2", "id": 4},
-                    {"name": "style3", "id": 6},
-                ],
-                "speaker_uuid": "7ffcb7ce-00ec-4bdc-82cd-45a8889e43ff",
-                "version": "mock",
-            },
-            {
-                "name": "dummy2",
-                "styles": [
-                    {"name": "style0", "id": 1},
-                    {"name": "style1", "id": 3},
-                    {"name": "style2", "id": 5},
-                    {"name": "style3", "id": 7},
-                ],
-                "speaker_uuid": "388f246b-8c41-4ac1-8e2d-5d79f3ff56d9",
-                "version": "mock",
-            },
-            {
-                "name": "dummy3",
-                "styles": [
-                    {"name": "style0", "id": 8},
-                ],
-                "speaker_uuid": "35b2c544-660e-401e-b503-0e14c635303a",
-                "version": "mock",
-            },
-            {
-                "name": "dummy4",
-                "styles": [
-                    {"name": "style0", "id": 9},
-                ],
-                "speaker_uuid": "b1a81618-b27b-40d2-b0ea-27a9ad408c4b",
-                "version": "mock",
-            },
-        ]
-    )
+                "cpu": True,
+                "cuda": False,
+            }
+        )
 
+    def finalize(self) -> None:
+        pass
 
-def supported_devices() -> str:
-    return json.dumps(
-        {
-            "cpu": True,
-            "cuda": False,
-        }
-    )
+    def load_model(self, style_id: int) -> None:
+        pass
+
+    def is_model_loaded(self, style_id: int) -> bool:
+        return True
+
+    def assert_core_success(self, result: bool) -> None:
+        pass
