@@ -9,13 +9,12 @@ from uuid import UUID, uuid4
 import numpy as np
 import pyopenjtalk
 from fastapi import HTTPException
-from pydantic import conint
 
 from .model import UserDictWord, WordTypes
 from .part_of_speech_data import MAX_PRIORITY, MIN_PRIORITY, part_of_speech_data
-from .utility import engine_root, get_save_dir, mutex_wrapper
+from .utility import get_save_dir, internal_root, mutex_wrapper
 
-root_dir = engine_root()
+root_dir = internal_root()
 save_dir = get_save_dir()
 
 if not save_dir.is_dir():
@@ -75,9 +74,11 @@ def update_dict(
         コンパイル済み辞書ファイルのパス
     """
     random_string = uuid4()
-    tmp_csv_path = save_dir / f".tmp.dict_csv-{random_string}"  # csv形式辞書データの一時保存ファイル
-    tmp_compiled_path = (
-        save_dir / f".tmp.dict_compiled-{random_string}"
+    tmp_csv_path = compiled_dict_path.with_suffix(
+        f".dict_csv-{random_string}.tmp"
+    )  # csv形式辞書データの一時保存ファイル
+    tmp_compiled_path = compiled_dict_path.with_suffix(
+        f".dict_compiled-{random_string}.tmp"
     )  # コンパイル済み辞書データの一時保存ファイル
 
     try:
@@ -434,16 +435,16 @@ def _search_cost_candidates(context_id: int) -> List[int]:
     raise HTTPException(status_code=422, detail="品詞IDが不正です")
 
 
-def _cost2priority(context_id: int, cost: conint(ge=-32768, le=32767)) -> int:
+def _cost2priority(context_id: int, cost: int) -> int:
+    assert -32768 <= cost <= 32767
     cost_candidates = _search_cost_candidates(context_id)
     # cost_candidatesの中にある値で最も近い値を元にpriorityを返す
     # 参考: https://qiita.com/Krypf/items/2eada91c37161d17621d
     # この関数とpriority2cost関数によって、辞書ファイルのcostを操作しても最も近いpriorityのcostに上書きされる
-    return MAX_PRIORITY - np.argmin(np.abs(np.array(cost_candidates) - cost))
+    return MAX_PRIORITY - np.argmin(np.abs(np.array(cost_candidates) - cost)).item()
 
 
-def _priority2cost(
-    context_id: int, priority: conint(ge=MIN_PRIORITY, le=MAX_PRIORITY)
-) -> int:
+def _priority2cost(context_id: int, priority: int) -> int:
+    assert MIN_PRIORITY <= priority <= MAX_PRIORITY
     cost_candidates = _search_cost_candidates(context_id)
     return cost_candidates[MAX_PRIORITY - priority]
